feat: per-channel distribution shape metrics + bf16 weight support
Add per-channel/per-group histogram slicing and distribution shape metrics (excess kurtosis, skewness, robust tail-ratio, shape label) with per-channel violin + shape-metrics heatmap visualization. Fix to_numpy to upcast bf16/fp16 -> fp32 (numpy has no bfloat16 dtype, so .numpy() on bf16 tensors raised).

+++ src/weiacviz/stats/_util.py
@@ -7,7 +7,9 @@ import numpy as np
 def to_numpy(t):
     """Convert torch.Tensor / np.ndarray / list to a float64 numpy array."""
     if hasattr(t, "detach"):
-        arr = t.detach().cpu().numpy()
+        # .float() upcasts bf16/fp16 -> fp32: numpy has no bfloat16 dtype, so
+        # calling .numpy() on a bf16 tensor raises "unsupported ScalarType".
+        arr = t.detach().cpu().float().numpy()
     elif hasattr(t, "numpy"):
         arr = t.numpy()
     else:
